feat : 학습자료 CRD 컴포넌트 추가

+++ src/main/kotlin/com/jipsa/balearn/database/learning_file/entity/LearningFileEntityVO.kt
@@ -13,7 +13,7 @@ data class LearningFIleInfoVO(
     val fileUrl: String,
 
     @Column(nullable = false)
-    val size: Double,
+    val size: Long,
 
     @Column(nullable = false)
     val type: String
